add alert and refactor custom controls
