+++ app/src/constants/targetModes.ts
@@ -1,0 +1,11 @@
+enum TargetMode {
+  DISABLED = 'DISABLED',
+  GENERIC = 'GENERIC',
+  WHATS_HERE = 'WHATS_HERE',
+  CUSTOM_LAYER = 'CUSTOM_LAYER',
+  ACTIVITY = 'ACTIVITY',
+  ACTIVITY_GEO_TRACK = 'ACTIVITY_GEO_TRACK',
+  TILE_CACHE = 'TILE_CACHE'
+}
+
+export default TargetMode;
